Update config.h to include TAPPING_TERM
Add TAPPING_TERM definition to config.h.

TAPPING_TERM defines the amount of time in MS that a "tap" is considered to be a "hold". This affects how long it takes for a key with multiple functions to recognize the "hold" bound function. The definition has been added here to expose it an make it easier for users to find and modify.